Review and maintenance before public release 2
Tags: [maintenance][refactor][test]
Compile: no tested
Tests: no tested

+++ src/noa/gpu/cuda/Types.h
@@ -74,9 +74,9 @@ namespace noa::cuda {
     }
 
     struct Limits {
-        static constexpr size_t WARP_SIZE = 32;
-        static constexpr size_t MAX_THREADS = 1024;
-        static constexpr size_t MAX_X_BLOCKS = (1U << 31) - 1U;
-        static constexpr size_t MAX_YZ_BLOCKS = 65535;
+        static constexpr uint WARP_SIZE = 32;
+        static constexpr uint MAX_THREADS = 1024;
+        static constexpr uint MAX_X_BLOCKS = (1U << 31) - 1U;
+        static constexpr uint MAX_YZ_BLOCKS = 65535;
     };
 }
